Extract some DPV classes to their own files
