+++ src/main/groovy/com/seriouslypro/pnpconvert/MaterialAssignment.groovy
@@ -1,0 +1,10 @@
+package com.seriouslypro.pnpconvert
+
+import groovy.transform.ToString
+
+@ToString(includeNames = true, includePackage = false)
+class MaterialAssignment {
+    Component component
+    Integer feederId
+    Feeder feeder
+}
