+++ src/main/groovy/com/seriouslypro/pnpconvert/MaterialSelectionEntry.groovy
@@ -1,0 +1,9 @@
+package com.seriouslypro.pnpconvert
+
+import groovy.transform.ToString
+
+@ToString(includeNames = true, includePackage = false)
+class MaterialSelectionEntry {
+    Component component
+    Feeder feeder
+}
